refactor: DTO 네이밍 수정 (#106)
