+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/post/application/PostDtoAssembler.java
@@ -2,7 +2,7 @@ package com.woowacourse.pickgit.post.application;
 
 import com.woowacourse.pickgit.authentication.domain.user.AppUser;
 import com.woowacourse.pickgit.post.application.dto.CommentResponse;
-import com.woowacourse.pickgit.post.application.dto.PostDto;
+import com.woowacourse.pickgit.post.application.dto.response.PostResponseDto;
 import com.woowacourse.pickgit.post.domain.Post;
 import com.woowacourse.pickgit.tag.domain.Tag;
 import com.woowacourse.pickgit.user.domain.User;
@@ -14,13 +14,13 @@ public class PostDtoAssembler {
     private PostDtoAssembler() {
     }
 
-    public static List<PostDto> assembleFrom(AppUser appUser, List<Post> posts) {
+    public static List<PostResponseDto> assembleFrom(AppUser appUser, List<Post> posts) {
         return posts.stream()
             .map(post -> convertFrom(post, appUser))
             .collect(Collectors.toList());
     }
 
-    private static PostDto convertFrom(Post post, AppUser appUser) {
+    private static PostResponseDto convertFrom(Post post, AppUser appUser) {
         User postWriter = post.getUser();
         List<String> tags = post.getTags()
             .stream()
@@ -32,13 +32,13 @@ public class PostDtoAssembler {
             .collect(Collectors.toList());
 
         if (appUser.isGuest()) {
-            return new PostDto(post.getId(), post.getImagaeUrls(), post.getGithubRepoUrl(),
+            return new PostResponseDto(post.getId(), post.getImagaeUrls(), post.getGithubRepoUrl(),
                 post.getContent(), postWriter.getName(), postWriter.getBasicProfile().getImage(),
                 post.getLikeCounts(),
                 tags, post.getCreatedAt(), post.getUpdatedAt(), comments, null);
         }
 
-        return new PostDto(post.getId(), post.getImagaeUrls(), post.getGithubRepoUrl(),
+        return new PostResponseDto(post.getId(), post.getImagaeUrls(), post.getGithubRepoUrl(),
             post.getContent(), postWriter.getName(), postWriter.getBasicProfile().getImage(),
             post.getLikeCounts(),
             tags, post.getCreatedAt(), post.getUpdatedAt(), comments, post.isLikedBy(appUser.getUsername()));
