+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/post/application/PostServiceIntegrationTest.java
@@ -12,10 +12,10 @@ import com.woowacourse.pickgit.exception.platform.PlatformHttpErrorException;
 import com.woowacourse.pickgit.exception.post.CommentFormatException;
 import com.woowacourse.pickgit.post.PostTestConfiguration;
 import com.woowacourse.pickgit.post.application.dto.CommentResponse;
-import com.woowacourse.pickgit.post.application.dto.PostDto;
+import com.woowacourse.pickgit.post.application.dto.response.PostResponseDto;
 import com.woowacourse.pickgit.post.application.dto.request.PostRequestDto;
 import com.woowacourse.pickgit.post.application.dto.request.RepositoryRequestDto;
-import com.woowacourse.pickgit.post.application.dto.response.PostResponseDto;
+import com.woowacourse.pickgit.post.application.dto.response.PostImageUrlResponseDto;
 import com.woowacourse.pickgit.post.application.dto.response.RepositoriesResponseDto;
 import com.woowacourse.pickgit.post.domain.Post;
 import com.woowacourse.pickgit.post.domain.PostRepository;
@@ -142,7 +142,7 @@ class PostServiceIntegrationTest {
                 ), githubRepoUrl, tags, content);
 
         // when
-        PostResponseDto responseDto = postService.write(requestDto);
+        PostImageUrlResponseDto responseDto = postService.write(requestDto);
 
         // then
         assertThat(responseDto.getId()).isNotNull();
@@ -198,17 +198,17 @@ class PostServiceIntegrationTest {
 
         HomeFeedRequest homeFeedRequest =
             new HomeFeedRequest(new LoginUser("kevin", "a"), 1L, 2L);
-        List<PostDto> postDtos = postService.readHomeFeed(homeFeedRequest);
+        List<PostResponseDto> postResponseDtos = postService.readHomeFeed(homeFeedRequest);
 
-        List<String> postNames = postDtos.stream()
-            .map(PostDto::getAuthorName)
+        List<String> postNames = postResponseDtos.stream()
+            .map(PostResponseDto::getAuthorName)
             .collect(toList());
 
-        List<String> repoNames = postDtos.stream()
-            .map(PostDto::getGithubRepoUrl)
+        List<String> repoNames = postResponseDtos.stream()
+            .map(PostResponseDto::getGithubRepoUrl)
             .collect(toList());
 
-        assertThat(postDtos).hasSize(2);
+        assertThat(postResponseDtos).hasSize(2);
         assertThat(postNames).containsExactly("ginger", "dani");
         assertThat(repoNames).containsExactly("jwp-chess", "java-racingcar");
     }
@@ -218,7 +218,7 @@ class PostServiceIntegrationTest {
         List<User> users = PostFactory.mockUsers();
         for (int i = 0; i < postRequestDtos.size(); i++) {
             userRepository.save(users.get(i));
-            PostResponseDto response = postService.write(postRequestDtos.get(i));
+            PostImageUrlResponseDto response = postService.write(postRequestDtos.get(i));
             CommentRequest commentRequest =
                 new CommentRequest(users.get(i).getName(), "test comment" + i, response.getId());
             postService.addComment(commentRequest);
@@ -242,13 +242,13 @@ class PostServiceIntegrationTest {
         //when
         HomeFeedRequest homeFeedRequest =
             new HomeFeedRequest(new LoginUser("kevin", "a"), 0L, 3L);
-        List<PostDto> postDtos = postService.readMyFeed(homeFeedRequest);
-        List<String> repoNames = postDtos.stream()
-            .map(PostDto::getGithubRepoUrl)
+        List<PostResponseDto> postResponseDtos = postService.readMyFeed(homeFeedRequest);
+        List<String> repoNames = postResponseDtos.stream()
+            .map(PostResponseDto::getGithubRepoUrl)
             .collect(toList());
 
         //then
-        assertThat(postDtos).hasSize(3);
+        assertThat(postResponseDtos).hasSize(3);
         assertThat(repoNames).containsExactly("java-racingcar", "jwp-chess", "atdd-subway-fare");
     }
 
@@ -270,16 +270,16 @@ class PostServiceIntegrationTest {
         //when
         HomeFeedRequest homeFeedRequest =
             new HomeFeedRequest(new LoginUser("ala", "a"), 0L, 3L);
-        List<PostDto> postDtos = postService.readUserFeed(homeFeedRequest, "kevin");
-        List<String> repoNames = postDtos.stream()
-            .map(PostDto::getGithubRepoUrl)
+        List<PostResponseDto> postResponseDtos = postService.readUserFeed(homeFeedRequest, "kevin");
+        List<String> repoNames = postResponseDtos.stream()
+            .map(PostResponseDto::getGithubRepoUrl)
             .collect(toList());
-        List<Boolean> likes = postDtos.stream()
-            .map(PostDto::getIsLiked)
+        List<Boolean> likes = postResponseDtos.stream()
+            .map(PostResponseDto::getIsLiked)
             .collect(toList());
 
         //then
-        assertThat(postDtos).hasSize(3);
+        assertThat(postResponseDtos).hasSize(3);
         assertThat(repoNames).containsExactly("java-racingcar", "jwp-chess", "atdd-subway-fare");
         assertThat(likes).containsExactly(Boolean.FALSE, Boolean.FALSE, Boolean.FALSE);
     }
@@ -302,16 +302,16 @@ class PostServiceIntegrationTest {
         //when
         HomeFeedRequest homeFeedRequest =
             new HomeFeedRequest(new GuestUser(), 0L, 3L);
-        List<PostDto> postDtos = postService.readUserFeed(homeFeedRequest, "kevin");
-        List<String> repoNames = postDtos.stream()
-            .map(PostDto::getGithubRepoUrl)
+        List<PostResponseDto> postResponseDtos = postService.readUserFeed(homeFeedRequest, "kevin");
+        List<String> repoNames = postResponseDtos.stream()
+            .map(PostResponseDto::getGithubRepoUrl)
             .collect(toList());
-        List<Boolean> likes = postDtos.stream()
-            .map(PostDto::getIsLiked)
+        List<Boolean> likes = postResponseDtos.stream()
+            .map(PostResponseDto::getIsLiked)
             .collect(toList());
 
         //then
-        assertThat(postDtos).hasSize(3);
+        assertThat(postResponseDtos).hasSize(3);
         assertThat(repoNames).containsExactly("java-racingcar", "jwp-chess", "atdd-subway-fare");
         assertThat(likes).containsExactly(null, null, null);
     }
